✨ add liftQubits / daggerGates method
To construct miter cirucit, we have to lift all the qubits in U by
nQubits. Also we have to dagger all the gates in U. (When applying
gates in U, we have to transpose gates, transposing gates are
already done in the API of applyGates.) We add the liftQubits /
daggerGates method to class Circuit and Gate to provide API for
construct miter.