✨ add simulation guide NEQ determination
For NEQ cases, conclusion can be obtained by simulating a few
basis state. More often, a single basic state simulation will
verify that the two circuit are NEQ.
In this way, we don't have to compute and check the whole
functionality, which may save a lot of runtime and momory.
I add checkBySimulation() function implementing this approach.